split webpack config up
