+++ tools/__tasks__/compile/javascript/webpack.dev.js
@@ -4,18 +4,18 @@ require('any-observable/register/rxjs-all');
 const Observable = require('any-observable');
 
 const webpack = require('webpack');
-const ProgressPlugin = require('webpack/lib/ProgressPlugin');
 const chalk = require('chalk');
+
+const config = require('../../../../webpack.config.dev.js');
 
 module.exports = {
     description: 'Create Webpack bundles',
     task: () => new Observable(observer => {
-        const config = require('../../../../webpack.config.js')({
-            plugins: [
-                new ProgressPlugin((progress, msg) =>
-                    observer.next(`${Math.round(progress * 100)}% ${msg}`)),
-            ],
-        });
+        config.plugins = [
+            require('../../../webpack-progress-reporter')(observer),
+            ...config.plugins,
+        ];
+
         const bundler = webpack(config);
 
         bundler.run((err, stats) => {
