+++ tools/__tasks__/compile/javascript/webpack.js
@@ -4,30 +4,18 @@ require('any-observable/register/rxjs-all');
 const Observable = require('any-observable');
 
 const webpack = require('webpack');
-const ProgressPlugin = require('webpack/lib/ProgressPlugin');
-const Visualizer = require('webpack-visualizer-plugin');
 const chalk = require('chalk');
+
+const config = require('../../../../webpack.config.prod.js');
 
 module.exports = {
     description: 'Create Webpack bundles',
     task: () => new Observable(observer => {
-        const config = require('../../../../webpack.config.js')({
-            env: 'production',
-            plugins: [
-                new webpack.optimize.AggressiveMergingPlugin(),
-                new Visualizer({
-                    filename: './webpack-stats.html',
-                }),
-                new webpack.DefinePlugin({
-                    'process.env.NODE_ENV': JSON.stringify('production'),
-                }),
-                new webpack.optimize.UglifyJsPlugin({
-                    sourceMap: true,
-                }),
-                new ProgressPlugin((progress, msg) =>
-                    observer.next(`${Math.round(progress * 100)}% ${msg}`)),
-            ],
-        });
+        config.plugins = [
+            require('../../../webpack-progress-reporter')(observer),
+            ...config.plugins,
+        ];
+
         const bundler = webpack(config);
 
         bundler.run((err, stats) => {
