+++ webpack.config.common.js
@@ -4,8 +4,10 @@ const webpack = require('webpack');
 
 const bundlePrefix = 'graun.';
 
-module.exports = ({ env = 'dev', plugins = [] } = {}) => ({
-    devtool: env === 'dev' ? 'inline-source-map' : 'source-map',
+module.exports.bundlePrefix = bundlePrefix;
+
+module.exports.config = {
+    devtool: 'source-map',
     entry: {
         standard: path.join(
             __dirname,
@@ -41,8 +43,8 @@ module.exports = ({ env = 'dev', plugins = [] } = {}) => ({
     },
     output: {
         path: path.join(__dirname, 'static', 'target', 'javascripts'),
-        filename: `${env === 'dev' ? '' : '[chunkhash]/'}${bundlePrefix}[name].js`,
-        chunkFilename: `${env === 'dev' ? '' : '[chunkhash]/'}${bundlePrefix}[name].js`,
+        filename: `[chunkhash]/${bundlePrefix}[name].js`,
+        chunkFilename: `[chunkhash]/${bundlePrefix}[name].js`,
     },
     resolve: {
         modules: [
@@ -94,10 +96,8 @@ module.exports = ({ env = 'dev', plugins = [] } = {}) => ({
         new webpack.ProvidePlugin({
             videojs: 'videojs',
         }),
-        // optional plugins passed in in production
-        ...plugins,
     ],
     externals: {
         xhr2: {},
     },
-});
+};
